Rename files to avoid confusion to pytest
