+++ examples/server_client/client_resilience.py
@@ -22,7 +22,7 @@ import os
 test = {
   "deployment_code": os.path.join(os.path.dirname(os.path.abspath(__file__)), "deployment.py"),
   "test_code": [
-      os.path.join(os.path.dirname(os.path.abspath(__file__)), "test_suites/test_resilience.py")],
+      os.path.join(os.path.dirname(os.path.abspath(__file__)), "test_suites/client_kill_recovery.py")],
   "perf_code": os.path.join(os.path.dirname(os.path.abspath(__file__)), "perf.py"),
   "configs_directory": os.path.join(os.path.dirname(os.path.abspath(__file__)), "configs/")
 }
